+++ server/impl/src/main/java/com/walmartlabs/concord/server/landing/LandingPageResourceImpl.java
@@ -13,11 +13,15 @@ import com.walmartlabs.concord.server.api.team.TeamEntry;
 import com.walmartlabs.concord.server.api.team.TeamRole;
 import com.walmartlabs.concord.server.project.ProjectDao;
 import com.walmartlabs.concord.server.project.ProjectManager;
+import com.walmartlabs.concord.server.project.RepositoryDao;
 import com.walmartlabs.concord.server.repository.RepositoryManager;
 import com.walmartlabs.concord.server.security.UserPrincipal;
 import com.walmartlabs.concord.server.team.TeamManager;
+import org.apache.shiro.authz.UnauthorizedException;
 import org.jooq.Configuration;
 import org.postgresql.util.Base64;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.sonatype.siesta.Resource;
 import org.sonatype.siesta.ValidationErrorsException;
 
@@ -35,24 +39,33 @@ import static com.walmartlabs.concord.server.project.RepositoryUtils.assertRepos
 @Named
 public class LandingPageResourceImpl extends AbstractDao implements LandingPageResource, Resource {
 
+    private static final Logger log = LoggerFactory.getLogger(LandingPageResourceImpl.class);
+
     private static final String LP_META_FILE_NAME = "landing.json";
 
     private final ObjectMapper objectMapper = new ObjectMapper();
     private final TeamManager teamManager;
     private final LandingDao landingDao;
     private final ProjectDao projectDao;
+    private final RepositoryDao repositoryDao;
     private final ProjectManager projectManager;
     private final RepositoryManager repositoryManager;
 
     @Inject
     public LandingPageResourceImpl(Configuration cfg,
-                                   TeamManager teamManager, LandingDao landingDao, ProjectDao projectDao,
-                                   ProjectManager projectManager, RepositoryManager repositoryManager) {
+                                   TeamManager teamManager,
+                                   LandingDao landingDao,
+                                   ProjectDao projectDao,
+                                   RepositoryDao repositoryDao,
+                                   ProjectManager projectManager,
+                                   RepositoryManager repositoryManager) {
+
         super(cfg);
 
         this.teamManager = teamManager;
         this.landingDao = landingDao;
         this.projectDao = projectDao;
+        this.repositoryDao = repositoryDao;
         this.projectManager = projectManager;
         this.repositoryManager = repositoryManager;
     }
@@ -105,23 +118,25 @@ public class LandingPageResourceImpl extends AbstractDao implements LandingPageR
     }
 
     @Override
+    public Response refreshAll() {
+        assertAdmin();
+        repositoryDao.list().parallelStream().forEach(r -> {
+            try {
+                refresh(r);
+            } catch (Exception e) {
+                log.warn("refreshAll -> {} refresh failed: {}", r.getId(), e.getMessage());
+            }
+        });
+        return Response.ok().build();
+    }
+
+    @Override
     public Response refresh(String projectName, String repositoryName) {
         UUID teamId = TeamManager.DEFAULT_TEAM_ID;
         ProjectEntry p = assertProject(teamId, projectName, TeamRole.WRITER, true);
         RepositoryEntry r = assertRepository(p, repositoryName);
 
-        Path lpMetaFile = repositoryManager.fetch(p.getId(), r)
-                .resolve(LP_META_FILE_NAME);
-
-        LandingEntry le = loadEntry(lpMetaFile);
-
-        tx(tx -> {
-            landingDao.delete(tx, p.getId(), r.getId());
-            if (le != null) {
-                byte[] icon = le.getIcon() != null ? Base64.decode(le.getIcon()) : null;
-                landingDao.insert(tx, p.getId(), r.getId(), le.getName(), le.getDescription(), icon);
-            }
-        });
+        refresh(r);
 
         return Response.ok().build();
     }
@@ -139,6 +154,23 @@ public class LandingPageResourceImpl extends AbstractDao implements LandingPageR
         }
     }
 
+    private void refresh(RepositoryEntry r) {
+        Path lpMetaFile = repositoryManager.fetch(r.getProjectId(), r)
+                .resolve(LP_META_FILE_NAME);
+
+        LandingEntry le = loadEntry(lpMetaFile);
+
+        tx(tx -> {
+            landingDao.delete(tx, r.getProjectId(), r.getId());
+            if (le != null) {
+                byte[] icon = le.getIcon() != null ? Base64.decode(le.getIcon()) : null;
+                landingDao.insert(tx, r.getProjectId(), r.getId(), le.getName(), le.getDescription(), icon);
+            }
+        });
+
+        log.info("refresh ['{}'] -> done", r.getId());
+    }
+
     private ProjectEntry assertProject(UUID teamId, String projectName, TeamRole role, boolean membersOnly) {
         if (projectName == null) {
             throw new ValidationErrorsException("A valid project name is required");
@@ -151,5 +183,12 @@ public class LandingPageResourceImpl extends AbstractDao implements LandingPageR
 
 
         return projectManager.assertProjectAccess(id, role, membersOnly);
+    }
+
+    private static void assertAdmin() {
+        UserPrincipal p = UserPrincipal.getCurrent();
+        if (!p.isAdmin()) {
+            throw new UnauthorizedException("Not authorized");
+        }
     }
 }